Renaming and more cleanup
