+++ lib/dragonrealms/drinfomon/drdefs.rb
