+++ lib/dragonrealms/drinfomon/drvariables.rb
